Modified x, y, z from double to float type

+++ Wireless_Sensor_Network/Disk.swift
@@ -12,7 +12,7 @@ public class Disk: RGG
 {
     public init(avgDegree: Int, numberOfVertices: Int)
     {
-        let r = sqrt(Double(avgDegree) / Double(numberOfVertices))
+        let r = sqrt(Float(avgDegree) / Float(numberOfVertices))
         super.init(r: r, numberOfVertices: numberOfVertices)        
         createVertices()
         createEdges()
@@ -24,12 +24,12 @@ public class Disk: RGG
     {
         for i in 0...(nVertices-1)
         {
-            var x:Double
-            var y:Double
+            var x:Float
+            var y:Float
             repeat
             {
-                x = Double.random(-1, max: 1)
-                y = Double.random(-1, max: 1)
+                x = Float.random(-1, max: 1)
+                y = Float.random(-1, max: 1)
             }while(square(x) + square(y) > 1)
             let v = Vertex(id: i, x: x, y: y)
             vertices.append(v)
